requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
feedparser==6.0.11
firecrawl-py==1.0.0
apscheduler==3.10.4
httpx==0.27.2
jinja2==3.1.4
